# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fix_Smart_CMS API",
    "version": "1.0.3",
    "description": "\n        Comprehensive API documentation for Fix_Smart_CMS (NLC-CMS) - A modern complaint management system.\n        \n        ## Features\n        - **Multi-role Authentication**: Citizens, Ward Officers, Maintenance Teams, Administrators\n        - **Complaint Management**: Complete lifecycle from registration to resolution\n        - **Geographic Routing**: Ward-based automatic complaint assignment\n        - **File Management**: Unified attachment system for all file types\n        - **Real-time Notifications**: Email and in-app notification system\n        - **Analytics & Reporting**: Comprehensive dashboard and reporting capabilities\n        \n        ## Authentication\n        Most endpoints require JWT authentication. Include the token in the Authorization header:\n        `Authorization: Bearer <your-jwt-token>`\n        \n        ## Rate Limiting\n        API requests are rate-limited to prevent abuse. Development environment has higher limits.\n        \n        ## Response Format\n        All API responses follow a consistent format:\n        ```json\n        {\n          \"success\": boolean,\n          \"message\": \"string\",\n          \"data\": object | array | null,\n          \"errors\": array (optional)\n        }\n        ```\n      ",
    "contact": {
      "name": "API Support",
      "email": "api-support@nlc-cms.gov.in",
...
    },
...
    },
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "pages": {
            "type": "integer",
            "example": 10
          },
          "hasNext": {
            "type": "boolean",
            "example": true
          },
          "hasPrev": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "UserManagement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "CITIZEN",
              "WARD_OFFICER",
              "MAINTENANCE_TEAM",
              "ADMINISTRATOR"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "ward": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number for pagination",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 10
        }
      },
      "SortParam": {
        "name": "sort",
        "in": "query",
        "description": "Sort field and direction (e.g., 'createdAt:desc')",
        "schema": {
          "type": "string",
          "default": "createdAt:desc"
        }
      },
      "SearchParam": {
        "name": "search",
        "in": "query",
        "description": "Search query string",
        "schema": {
          "type": "string"
        }
      },
      "StatusParam": {
        "name": "status",
        "in": "query",
        "description": "Filter by status",
        "schema": {
          "type": "string",
          "enum": [
            "REGISTERED",
            "ASSIGNED",
            "IN_PROGRESS",
            "RESOLVED",
            "CLOSED",
            "REOPENED"
          ]
        }
      },
      "PriorityParam": {
        "name": "priority",
        "in": "query",
        "description": "Filter by priority",
        "schema": {
          "type": "string",
          "enum": [
            "LOW",
            "MEDIUM",
            "HIGH",
            "CRITICAL"
          ]
        }
      },
      "WardParam": {
        "name": "ward",
        "in": "query",
        "description": "Filter by ward ID",
        "schema": {
          "type": "string"
        }
      },
      "DateFromParam": {
        "name": "dateFrom",
        "in": "query",
        "description": "Filter from date (ISO format)",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "DateToParam": {
        "name": "dateTo",
        "in": "query",
        "description": "Filter to date (ISO format)",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required. Please provide a valid token."
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/api/admin/users": {
      "get": {
        "summary": "Get all users (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ward",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserManagement"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "pages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "email",
                  "role"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "CITIZEN",
                      "WARD_OFFICER",
                      "MAINTENANCE_TEAM",
                      "ADMINISTRATOR"
                    ]
                  },
                  "wardId": {
                    "type": "string"
                  },
                  "department": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/complaints": {
      "get": {
        "summary": "Get complaints with filtering and pagination",
        "tags": [
          "Complaints"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/SortParam"
          },
          {
            "$ref": "#/components/parameters/SearchParam"
          },
          {
            "$ref": "#/components/parameters/StatusParam"
          },
          {
            "$ref": "#/components/parameters/PriorityParam"
          },
          {
            "$ref": "#/components/parameters/WardParam"
          },
          {
            "$ref": "#/components/parameters/DateFromParam"
          },
          {
            "$ref": "#/components/parameters/DateToParam"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by complaint type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "description": "Filter by assigned user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complaints retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "complaints": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Complaint"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMeta"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "summary": "Create a new complaint",
        "tags": [
          "Complaints"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "description",
                  "wardId",
                  "area",
                  "contactPhone"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Street light not working"
                  },
                  "description": {
                    "type": "string",
                    "example": "The street light on MG Road has been non-functional for 3 days"
                  },
                  "type": {
                    "type": "string",
                    "example": "STREET_LIGHTING"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "LOW",
                      "MEDIUM",
                      "HIGH",
                      "CRITICAL"
                    ],
                    "default": "MEDIUM"
                  },
                  "wardId": {
                    "type": "string",
                    "example": "ward123"
                  },
                  "subZoneId": {
                    "type": "string",
                    "example": "subzone123"
                  },
                  "area": {
                    "type": "string",
                    "example": "MG Road"
                  },
                  "landmark": {
                    "type": "string",
                    "example": "Near City Mall"
                  },
                  "address": {
                    "type": "string",
                    "example": "MG Road, Near City Mall, Kochi"
                  },
                  "latitude": {
                    "type": "number",
                    "example": 9.9312
                  },
                  "longitude": {
                    "type": "number",
                    "example": 76.2673
                  },
                  "contactName": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "contactEmail": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@example.com"
                  },
                  "contactPhone": {
                    "type": "string",
                    "example": "+91-9876543210"
                  },
                  "isAnonymous": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Complaint created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "complaint": {
                          "$ref": "#/components/schemas/Complaint"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/api/complaint-types": {
      "get": {
        "summary": "Get all complaint types",
        "tags": [
          "Complaint Types"
        ],
        "responses": {
          "200": {
            "description": "List of all complaint types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ComplaintType"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new complaint type",
        "tags": [
          "Complaint Types"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintType"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Complaint type created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/logs": {
      "post": {
        "summary": "Submit frontend logs",
        "tags": [
          "Logs"
        ],
        "description": "Submit client-side logs to the server for monitoring and debugging",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/LogEntry"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/LogEntry"
                    }
                  }
                ]
              },
              "examples": {
                "singleLog": {
                  "summary": "Single log entry",
                  "value": {
                    "level": "error",
                    "message": "Failed to submit complaint",
                    "timestamp": "2024-12-01T10:30:00Z",
                    "url": "/complaints/new",
                    "stack": "Error: Network request failed..."
                  }
                },
                "multipleLogs": {
                  "summary": "Multiple log entries",
                  "value": [
                    {
                      "level": "info",
                      "message": "User logged in",
                      "timestamp": "2024-12-01T10:00:00Z"
                    },
                    {
                      "level": "error",
                      "message": "API request failed",
                      "timestamp": "2024-12-01T10:30:00Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logs received successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "received": {
                          "type": "integer",
                          "description": "Number of log entries received"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid log format"
          },
          "429": {
            "description": "Too many log submissions (rate limited)"
          }
        }
      }
    },
...
    "/api/config/admin": {
      "get": {
        "summary": "Get all system configuration (Admin only)",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieve all system configuration values including sensitive data",
        "responses": {
          "200": {
            "description": "Admin configuration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cached": {
                          "type": "object",
                          "description": "Cached configuration values"
                        },
                        "database": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "key": {
                                "type": "string"
                              },
                              "value": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "isActive": {
                                "type": "boolean"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        },
                        "cacheStats": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
    "/api/config": {
      "post": {
        "summary": "Create new system configuration",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfigCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Configuration created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemConfig"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "409": {
            "description": "Configuration key already exists"
          }
        }
      }
    },
...
  }
}